PT:VZW:CV MMS not able to be sent with messenger app
Bug 17506557

On Sprint there's a case where the mcc/mnc was getting updated in the
subscription record. As an emergency backup, particularly on single-sim
devices on L, use the main context's mcc/mnc when the one in the
SubInfoRecord has not been set. Also add more default logging. This extra
logging only shows up at boot-up and an extra line when an MmsConfig is
requested from the MmsConfigManager.


Change-Id: Ie50cf08924be34f56aa13b8c382ae71d9e13caf7