Do not crash when mms config is not loaded when sending mms
In some case, it takes the system a while to return active subscriptions
which are required for mms service to load mms config. This would cause
a crash if we send mms before mms config is loaded. This fix makes sure
we don't crash and if that happens, gentally returns an error code to
the mms sender.

b/17556258

Change-Id: Ie180fb214b346e189c9ecae0d3bc4b15e4207fb1