PduParser MSIM support (1/4)
PduParser needs to know if Content-Disposition part header is supported
for a specific carrier. Originally this is loaded directly from system
resource in PduParser. This won't work for multi-SIM, which requires
getting that config based on carrier. This change adds parameter to
indicate if this header is supported and requires caller to provider the
config.

b/18371228

Change-Id: Ief4ef8942f5dce5251429887f312dcadb03dbf39