Catch exception for ConnectivityManager.unregisterNetworkCallback
It is possible ConnectivityManager.requestNetwork may fail silently due
to RemoteException. When that happens, we may get an invalid
NetworkCallback, which causes an IllegalArgumentexception when we try to
unregisterNetworkCallback. This exception in turn causes
MmsNetworkManager to skip the code to clean up states. Thus MMS service
would get stuck in the bad state until the device restarts. This fix
catches the exception so that state clean up can be executed.

b/24306554

Change-Id: I3c0df8dab9af0f7a7f2e647b38f5df29177b7ff8