MmsRequest: Switch data subscription on MSIM if needed
If the active data subscription isn't the one we're trying to send
the message out of, request a switch so that we don't try using
the device's data sub instead of the one that is needed.

In these scenarios, extend the retry cycle, too. We don't want
to flip back and forth when radio conditions are poor, so try
to get more mileage out of the temporary switch periods.

Ref CYNGNOS-2492

Change-Id: Ide3bcc65beb171c8db1f49f9c101a5db7fd505f4
(cherry picked from commit b7a3f51a512cd425059503c0cbb818f15f4291db)