MmsRequest: Reimplement automatic data sub switch
Make the whole process more synchronous, by waiting for a switch
confirmation from the subscription controller instead of waiting
for a fixed "safe" amount of time.

Also, tie the switches to the actual network requests/releases so
that the existing "reuse this connection if possible" hooks affect
the switches too. We want as few of those as possible.

This replaces the implementation from Ide3bcc65beb171c8db1f49f9c101a5db7fd505f4

Ref CYNGNOS-2492

Change-Id: Ic9bf7897150ae53e6ebb55c2e8fceb003a366417