Bump up MMS network request timeout and don't retry
Moving forward, the MMS service will rely on the retry logic of the
lower level Telephony layer, and retire it's own retry logic. Also, we
increase the timeout for network requests, to ensure we give enough time
for the Telephony retries to be attempted.

Test: did a manual test in which I turned airplane mode off, observed
that prior to the change, we tried 3 times, and after the change, only
once. Also observed that when connection is successful, it works Ok the
first time around (i.e. no regression).

Bug: 31383107

Change-Id: If4d67ec0731767e0ac55476f592b0bd9f6657955